+++ test/unit/fuzz-tools.test.js
@@ -486,6 +486,7 @@ import * as vhi10V508 from '../../lib/vhi10-v508.js';
 import * as sunnybrookFacialV509 from '../../lib/sunnybrook-facial-v509.js';
 import * as banffTcmrV510 from '../../lib/banff-tcmr-v510.js';
 import * as crafftV511 from '../../lib/crafft-v511.js';
+import * as vaizeyV512 from '../../lib/vaizey-v512.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -917,6 +918,7 @@ const MODULES = {
   'sunnybrook-facial-v509.js': sunnybrookFacialV509,
   'banff-tcmr-v510.js': banffTcmrV510,
   'crafft-v511.js': crafftV511,
+  'vaizey-v512.js': vaizeyV512,
   'tb-testing.js': tbTesting,
 };